requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.4
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
google-crc32c==1.7.1
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.5.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
onnxruntime==1.22.0
openai==1.70.0
opencv-python==4.11.0.86
orjson==3.10.18
packaging==25.0
pillow==11.1.0
pipecat-ai==0.0.71
pipecat-ai-small-webrtc-prebuilt==0.7.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
protobuf==5.29.5
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.5
pydantic-settings==2.10.0
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==0.12.0
pyloudnorm==0.1.1
pyOpenSSL==25.1.0
pyreadline3==3.5.4
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.4.3
resampy==0.4.3
rich==14.0.0
rich-toolkit==0.14.7
scipy==1.15.3
shellingham==1.5.4
silero-vad==5.1.2
six==1.17.0
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
sympy==1.14.0
tomli==2.2.1
# torch==2.7.1
# torchaudio==2.7.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==13.1
win32_setctime==1.2.0
yarl==1.20.1
# xformers==0.0.30
azure-cognitiveservices-speech==1.42.0
